Create Github release at the end of the release pipeline (#84)
Releases will be automatically created at the end of the release pipeline. Note that one extra field named `notesStartTag` should be provided as part of the `release.yml` file. This field will be used to specify which is the tag to be used to generated the commits diff.

```yaml
release:
  branch: develop
  version: 7.8.0
  nextVersion: 7.7.0-SNAPSHOT
  notesStartTag: 7.7.0
  mock: false
  baseTag:
    activiti: 7.7.0-alpha.13
    activitiCloud: 7.8.0-alpha.3
    activitiCloudApplication: 7.8.0-alpha.3
    commonChart: 7.7.0
    fullChart: 7.8.0-alpha.3
```

Ref: Activiti/Activiti#4189